Chore(test): Move juju version check
