+++ tests/integration/integrations/conftest.py
@@ -15,33 +15,12 @@ import pytest
 import pytest_asyncio
 from juju.application import Application
 from minio import Minio
-from ops import JujuVersion
 from pytest_operator.plugin import OpsTest
 
 PROJECT_ROOT = pathlib.Path(__file__).parent.parent.parent.parent
 nest_asyncio.apply()
 
 logger = logging.getLogger(__name__)
-
-
-@pytest.fixture(autouse=True)
-def skip_by_juju_version(request, model):
-    """Skip the test if juju version is lower then the `skip_juju_version` marker value."""
-    if request.node.get_closest_marker("skip_juju_version"):
-        current_version = JujuVersion(
-            f"{model.info.agent_version.major}.{model.info.agent_version.minor}.{model.info.agent_version.patch}"
-        )
-        min_version = JujuVersion(request.node.get_closest_marker("skip_juju_version").args[0])
-        if current_version < min_version:
-            pytest.skip("Juju version is too old")
-
-
-def pytest_configure(config):
-    """Add new marker."""
-    config.addinivalue_line(
-        "markers",
-        "skip_juju_version(version): skip test if Juju version is lower than version",
-    )
 
 
 @pytest.fixture(autouse=True)
